# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estore API",
        "description": "Estore Api",
        "contact": {
            "name": "Victor Maamaa",
            "email": "maaamaavictor@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/{product}/variations": {
            "get": {
                "tags": [
                    "Product Variations"
                ],
                "summary": "Get variations for a specific product",
                "operationId": "96adc54ecf78ce6d2a2a8159b388daa9",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Product variations retrieved successfully",
                                    "data": []
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product Variations"
                ],
                "summary": "Create a new variation for a specific product",
                "operationId": "730d4cb8ef309a3afa7f9bb561ba1703",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "weight_unit": {
                                        "type": "string"
                                    },
                                    "weight_value": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "height_unit": {
                                        "type": "string"
                                    },
                                    "height_value": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "width_unit": {
                                        "type": "string"
                                    },
                                    "width_value": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "depth_unit": {
                                        "type": "string"
                                    },
                                    "depth_value": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "volume_unit": {
                                        "type": "string"
                                    },
                                    "volume_value": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Product variation created successfully",
                                    "data": []
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{product}/variations/{variation}": {
            "get": {
                "tags": [
                    "Product Variations"
                ],
                "summary": "Get details of a specific product variation",
                "operationId": "239e61d297b523828a309fd82ff7fbeb",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variation",
                        "in": "path",
                        "description": "Product Variation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search parameters for product variation",
                        "required": false,
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "field1": "value1",
                                    "field2": "value2"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Product variation retrieved successfully",
                                    "data": []
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Product Variations"
                ],
                "summary": "Update details of a specific product variation",
                "operationId": "806dd2cf840423d34f4315c123fcf18b",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variation",
                        "in": "path",
                        "description": "Product Variation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Product Variation Name",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product Variation Price",
                                        "type": "number"
                                    },
                                    "weight": {
                                        "description": "Product Variation Weight",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Product variation updated successfully",
                                    "data": []
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product Variations"
                ],
                "summary": "Delete a specific product variation",
                "operationId": "c2d1bb21df6f09ada2373769a9927542",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variation",
                        "in": "path",
                        "description": "Product Variation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Product variation Deleted successfully",
                                    "data": []
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}